Fix TiCDC build on FreeBSD
